pyproject.toml
[project]
name = "my-lib"
version = "0.2.0"
description = "Private utility library"
authors = [
    { name = "KIMATA Tetsuya", email = "kimata@green-rabbit.net" }
]

readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "lxml>=5.3.0",
    "psutil>=6.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "slack-sdk>=3.31.0",
    "uptime>=3.0.1",
    "coloredlogs>=15.0.1",
    "docopt-ng>=0.9.0",
    "pyserial>=3.5",
    "smbus2>=0.4.3",
    "genson>=1.3.0",
    "jsonschema>=4.23.0",
    "line-bot-sdk>=3.14.1",
    "gitpython>=3.1.44",
    "pytz>=2025.2",
    "influxdb-client[ciso]>=1.44.0",
    "rich>=14.0.0",
    "sseclient>=0.0.27",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "flaky>=3.8.1",
    "playwright>=1.57.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.15.1",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.8.0",
    "pytest-timeout>=2.4.0",
    "time-machine>=3.2.0",
    "mypy>=1.19.1",
    "pyright>=1.1.407",
    "types-pytz>=2025.2.0",
    "types-paramiko>=4.0.0",
    "types-pyyaml>=6.0.12",
    "amazon-paapi5>=1.1.2",
    "flask-cors>=4.0.1",
    "flask>=3.0.3",
    "openpyxl>=3.1.5",
    "pillow>=10.4.0",
    "rpi-lgpio>=0.6",
    "selenium>=4.23.1",
    "spidev>=3.6",
    "speechrecognition>=3.11.0",
    "pydub>=0.25.1",
    "pytest-forked>=1.6.0",
    "pytest-rerunfailures>=15.1",
    "undetected-chromedriver>=3.5.5",
    "types-requests>=2.32.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/my_lib"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
minversion = "6.0"
# E2E テストはデフォルトで除外（外部サーバーが必要なため）
# E2E テスト実行: uv run pytest tests/e2e/ --host <host> --port <port>
addopts = "--verbose --timeout=300 --durations=10 --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys --junit-xml=reports/junit-report.xml --html=reports/pytest.html --self-contained-html --cov=src --cov-report=html --cov-report=term-missing --cov-report=xml:reports/coverage.xml --cov-report=lcov:reports/coverage.lcov --numprocesses=auto --ignore=tests/e2e"
testpaths = ["tests"]
filterwarnings = [
    "ignore:datetime\\.datetime\\.utcfromtimestamp\\(\\) is deprecated:DeprecationWarning",
    "ignore::DeprecationWarning:multiprocessing\\.popen_fork",
    # pydub の警告（Python 3.12+ でのエスケープシーケンスと audioop 非推奨）
    "ignore:invalid escape sequence:SyntaxWarning:pydub.*",
    "ignore:'audioop' is deprecated:DeprecationWarning:pydub.*",
    "ignore:unclosed database.*:ResourceWarning:rich.*",
    "ignore:unclosed database.*:ResourceWarning:influxdb_client.*",
    "ignore:unclosed database.*:ResourceWarning:coverage.*",
    "ignore:unclosed database.*:ResourceWarning:time_machine.*",
    "ignore:unclosed database.*:ResourceWarning:werkzeug.*",
    "ignore:unclosed database.*:ResourceWarning:_pytest.*",
    "ignore:unclosed database.*:ResourceWarning:typing.*",
    "ignore:unclosed database.*:ResourceWarning:multiprocessing.*",
    "ignore:unclosed database.*:ResourceWarning:logging.*",
    "ignore::ResourceWarning:_pytest.unraisableexception",
    "ignore::ResourceWarning:coverage.sqldata",
    "ignore::ResourceWarning:coverage.collector",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "reports/coverage"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_return_any = false
warn_unused_ignores = false
check_untyped_defs = true
no_implicit_optional = true
exclude = [
    "^tests/",
    "^.venv/",
]
warn_unused_configs = true

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
include = ["src", "tests"]
exclude = ["tests/fixtures"]
reportMissingImports = false
reportArgumentType = false
reportOperatorIssue = false
reportIndexIssue = false
reportGeneralTypeIssues = false
